Event Scrollbar, event icon fix
